Added setup-hlf github action
Hello, community!
I wish to add my GitHub action "setup-hlf" to the repository. I wanted to set up CI/CD pipelines through GitHub actions for a hlf project I was working on and I felt the need for an action to set hlf on the runner easily. There are still many more features to add but I hope it will help and benefit the hlf community.